1.ARM端添加101平衡模式程序；

+++ RN2012-F30-DSP/host/readconfig.h
@@ -23,7 +23,7 @@ struct _YCOVERPARAME_
 {
 	short ycindex;				//遥测编号
 	short flag;					//越限标志 上限1  下限2
-	int value;				    //越限值
+	short value;				//越限值
 };
 typedef struct _YCOVERPARAME_ YCOVERSTRM;
